[clang-tidy] ChangeLevel: Change argument type to match use

+++ src/scene/ChangeLevel.h
@@ -51,7 +51,7 @@ ZeniMax Media Inc., Suite 120, Rockville, Maryland 20850 USA.
 
 namespace fs { class path; }
 
-void ARX_CHANGELEVEL_Change(const std::string & level, const std::string & target, long angle);
+void ARX_CHANGELEVEL_Change(const std::string & level, const std::string & target, float angle);
 
 long ARX_CHANGELEVEL_GetInfo(const fs::path & savefile, std::string & name, float & version, long & level);
